Set _imageUrl to null in BowlerAddEditViewModel
The _imageUrl field in the BowlerAddEditViewModel class was originally initialized to an empty string (string.Empty). This change updates the initialization to null, which may be more appropriate for representing the absence of an image URL.

This was breaking change which prevented the default person.png from showing on an Add Bowler screen.