Allow nullable ImageUrl in person-related records
Updated the `PersonImageAddedOrChangedMessage` record to accept a nullable `ImageUrl`.
Modified the `ImageUrl` property in the `Person` class to be nullable and initialized to `null`.
Adjusted the mapping in `WeekMapper.cs` to use the nullable `ImageUrl` directly.
Changed the constructor of `CurrentWeekListItemViewModel` to accept a nullable `imageUrl` parameter.

This was a bug with ImageUrl not being null and therefore, preventing the initials (text) from showing in the avatar