Add "BLIND" label support in UserImage component
Introduced a new bindable property `ShowBlind` to control the visibility of a "BLIND" label in the `UserImage.xaml` file. The label is displayed within a border and is conditionally visible based on the `ShowBlind` property bound to the `AvatarView`. Adjusted the position of the existing `ShowSub` property for clarity. The `ShowBlindProperty` is defined as a static readonly bindable property with a default value of `false`.